Created component for sidebar with filter
