+++ src/app/[locale]/(logged-out)/data-custodian-network/[dataCustodianNetworkId]/components/DataCustodianNetwork/DataCustodianNetwork.tsx
@@ -2,10 +2,9 @@
 
 import HeaderActionBar from "@/components/HeaderActionBar";
 import LayoutDataItemPage from "@/components/LayoutDataItemPage";
-import ActiveListSidebar from "@/modules/ActiveListSidebar";
 import { Box, Typography } from "@mui/material";
 import IntroductionContent from "../IntroductionContent";
-import { Suspense, useEffect, useMemo, useState } from "react";
+import { Suspense, useMemo, useState } from "react";
 import { NetworkSkeleton, SectionSkeleton } from "@/components/Skeletons";
 import DataCustodianOuter from "../DataCustodianOuter";
 import DatasetsOuter from "../DatasetsOuter";
@@ -21,6 +20,7 @@ import { Filter, FilterValues } from "@/interfaces/Filter";
 import { FILTER_PUBLISHER_NAME } from "@/config/forms/filters";
 import { DatasetsSummaryData, EntitiesSummaryData, NetworkCustodiansSummaryData } from "@/interfaces/DataCustodianNetwork";
 import { SearchCategory } from "@/interfaces/Search";
+import ActiveListSidebarWithFilter from "@/modules/ActiveListSidebarWithFilter";
 
 const TRANSLATION_PATH = "pages.dataCustodianNetwork";
 
@@ -72,7 +72,7 @@ export default function DataCustodianNetwork({infoData, dataNetworkCustodiansSum
 
     return (
         <LayoutDataItemPage
-            navigation={<ActiveListSidebar items={activeLinkList} filter={publisherFilter} filterValues={filterValues} onFilterChange={setFilterValues} />}
+            navigation={<ActiveListSidebarWithFilter items={activeLinkList} filter={publisherFilter} filterValues={filterValues} onFilterChange={setFilterValues} />}
             body={
                 <>
                     <Typography variant="h1" sx={{ ml: 2, mt: 2 }}>
