+++ src/modules/ActiveListSidebarWithFilter/index.tsx
@@ -1,0 +1,3 @@
+import ActiveListSidebarWithFilter from "./ActiveListSidebarWithFilter";
+
+export default ActiveListSidebarWithFilter;
